feat: Created orders integration.

+++ HandleService/app/src/main/java/br/com/handleservice/domain/model/Service.kt
@@ -2,10 +2,10 @@ package br.com.handleservice.domain.model
 
 import java.time.LocalTime
 
-data class Service (
+data class Service(
     val id: Int,
     val enable: Boolean,
-    val value: Int,
+    val value: Double,
     val name: String,
     val estimatedTime: LocalTime,
     val description: String,
